+++ src/main/java/dev/doctor4t/trainmurdermystery/util/GunShootPayload.java
@@ -57,6 +57,12 @@ public record GunShootPayload(int target) implements CustomPayload {
                 var game = GameWorldComponent.KEY.get(player.getWorld());
                 Item revolver = TMMItems.REVOLVER;
                 if (game.isInnocent(target) && !player.isCreative() && mainHandStack.isOf(revolver)) {
+                    // backfire: if you kill an innocent you have a chance of shooting yourself instead
+                    if (player.getRandom().nextFloat() <= game.getBackfireChance()) {
+                        GameFunctions.killPlayer(player, true, player, TMM.id("gun_shot"));
+                        return;
+                    }
+
                     Scheduler.schedule(() -> {
                         if (!context.player().getInventory().contains((s) -> s.isIn(TMMItemTags.GUNS))) return;
                         player.getInventory().remove((s) -> s.isOf(revolver), 1, player.getInventory());